fixed an alert as a first step before the modal

+++ script.js
@@ -7,6 +7,7 @@ const button = document.querySelector(".btn");
 
 
 
+
 //SKAPAR EN TOM LOKAL ARRAY
 todosArray = []
 
@@ -58,20 +59,30 @@ const getTodos = async () => {
     const button = document.createElement("button")
     button.classList.add("btn")
     button.innerText = "delete"
-    button.addEventListener("click", ( )=>{
-      fetch(BASE_URL + todoData.id, {
+
+
+
+    
+    button.addEventListener("click", ( ) => {
+      // Deletes a Todo
+      if (!checkbox.checked) {
+        alert("Please check the checkbox before proceeding.");
+        return; }
+        fetch(BASE_URL + todoData.id, {
         method: 'DELETE'
-      })
+        }) 
+
         .then(res => {
-          // console.log(res)
+          console.log(res)
           if(res.ok) {
             li.remove()
             const index = (todosArray.findIndex(todo => todo.id == todoData.id))
             todosArray.splice(index, 1)
-            // console.log(todosArray)
+            console.log(todosArray)
           }
         })
-    } )
+      
+    })
 
     const checkbox = document.createElement("input")
     if(todoData.completed){
@@ -81,6 +92,7 @@ const getTodos = async () => {
     checkbox.style.width = "30px";
     checkbox.style.height = "30px";
     checkbox.style.cursor = "pointer";
+
     checkbox.addEventListener("change", () => {
       fetch(BASE_URL + todoData.id, {
         method: "PATCH",
@@ -119,30 +131,6 @@ const getTodos = async () => {
     return li
   }
 
-//   const removeUser = e => {
-  
-//   if(!e.target.classList.contains("header")) {
-//     console.log("Du klickade inte på headern")
-//     return
-//   }
- 
-//   fetch(BASE_URL + e.target.id, {
-//     method: 'DELETE'
-//   })
-//     .then(res => {
-//       // console.log(res)
-//       if(res.ok) {
-//         e.target.remove()
-//         const index = (todosArray.findIndex(todo => todo.id == e.target.id))
-//         todosArray.splice(index, 1)
-//         // console.log(todosArray)
-//       }
-//     })
-// }
-  
-
- 
-  // ul.addEventListener("click", removeUser)
